Add optional X-Atat-Api-Version header
As the ATAT API specification is modified ahead of a v1.0.0, it may be
helpful for client implementations to indicate the specific version that
they support and are expecting for the response. This new header is
entirely optional and purely informative.

Alternative solution here may have included:

* Leveraging the `Accept` headers -- this was not done because it would
  be a bit awkward to properly represent in the OpenAPI specification as
  we'd want to define the specific models that we expect in the
  response for the different headers. It would also suggest that we'd
  want a 406 Not Acceptable response if the server didn't support the
  given version; at this stage, we likely want to have everyone do their
  best to use and pass data around. Perhaps at/after stabilization the
  goals might be different.
* Expecting implementations to use path-based versioning -- this works
  for discrete versioning like /api/v1 or /api/v2, but gets awkward for
  pre-v1. Additionally, it implies too much burden to manually
  understand what each implementor supports.
* Expecting that everyone supports the latest version -- this basically
  what we were doing before. There was uncertainty about when a
  particular version was expected; at least with this there is a
  concrete way for partners to monitor what version of the API a client
  is expecting; some implementations may choose to support multiple
  versions and return a response compatible with an older version.

The downsides primarily are that this adds an additional header and that
it doesn't provide any help if a server implementation does not yet
support a new path; however, at least the client would potentially
include this header.

Clients really should include this header but it entirely optional and
not marked as being a "should do" to avoid anyone relying on this header
or assigning too much meaning to it.

In review, please ensure that the wording of the description meets the
intent and that there aren't typos or other mistakes.