Use a single object for CostData-type responses
The previous model allowed for an array of `CostData` objects in most
situations. This resulted in a structure where it's unclear why there
would be separation into multiple arrays and when/how to do so. Instead,
for the given date range, all results should be included in a single
`CostData` object and the months should be within the `results` array
inside the `CostData` object.

The `CostResponsebyClin` example already was using this single-item
format in its example; the `CostResponseByPortfolio` examples have been
updated to match the new specification.