Define the endpoint for querying provisioning status
The current specification provides an implied way for this to be
implemented but isn't overly opinionated on the matter. The current
specification doesn't make it obvious whether futher requests should
include the `Location` header in the response and whether these should
be queryable permanently.

This draft wording requires that provisioning status is only
_temporarily_ queryable but that IDs remain unique. The time period for
which a provisioning job can be queried is 7-10 days (specified in
hours). Additionally, it is clarified that the HTTP status should always
be a 200 and that it should return a ProvisioningStatus object but
returning the Location header is not required. The specification
requiring unique IDs (even across portfolios) matches the recommended
path structure in the previous example and may match potential
implementations under the existing specification.

Keeping provisoning jobs only queryable temporarily means that client
implementations will be forced to eventually recognize errors in their
structure (by handling the 404 response) rather than eternally looping
and querying status. Rather than requiring the expiration occur 168-240
hours after being created, we may want to consider requiring that after
changing from the `IN_PROGRESS` state.